Eliminated the HeliocentricCoordinates class.  All coordinate transformations will now just be top level functions in CoordinateManipulations.
