+++ app/src/main/java/com/google/android/stardroid/space/Moon.kt
@@ -1,7 +1,6 @@
 package com.google.android.stardroid.space
 
 import com.google.android.stardroid.ephemeris.Planet
-import com.google.android.stardroid.math.HeliocentricCoordinates
 import com.google.android.stardroid.math.RaDec
 import java.util.*
