+++ app/src/main/java/com/google/android/stardroid/space/SunOrbitingObject.kt
@@ -1,9 +1,9 @@
 package com.google.android.stardroid.space
 
 import com.google.android.stardroid.ephemeris.Planet
-import com.google.android.stardroid.math.HeliocentricCoordinates
-import com.google.android.stardroid.math.HeliocentricCoordinates.Companion.heliocentricCoordinatesFromOrbitalElements
 import com.google.android.stardroid.math.RaDec
+import com.google.android.stardroid.math.convertToEquatorialCoordinates
+import com.google.android.stardroid.math.heliocentricCoordinatesFromOrbitalElements
 import java.util.*
 
 /**
@@ -14,7 +14,7 @@ open class SunOrbitingObject(private val planet : Planet) : SolarSystemObject() 
         val earthCoords = getEarthHeliocentricCoordinates(date)
         var myCoords = getMyHeliocentricCoordinates(date)
         myCoords.minusAssign(earthCoords)
-        val equ = HeliocentricCoordinates.convertToEquatorialCoordinates(myCoords)
+        val equ = convertToEquatorialCoordinates(myCoords)
         return RaDec.calculateRaDecDist(equ)
     }
